[OpenXR-Loader] Add support for loading libraries from zip files
The linker from bionic on Android supports loading libraries from zip files.
This is done by using a path of the format of /path/to/zip!/path/to/lib
Currently if such a path is attempted to be loaded it will fail when it tries
to check if that path exists not knowing that this is in a special format.
Currently when faced with such a path it thinks that the library does not exit
and will abort trying to use it.

This change adds limited support for library paths of this format.
It currently only checks for the existence of the zip file and assumes that the
library within exists. This avoids the complexity of trying to list the entries
of the zip file.

This functionality is needed to support the library path found on v62 which is
/system/priv-app/VrDriver/VrDriver.apk!/lib/arm64-v8a/libopenxr_forwardloader.so